Replace duplicate CHUNK_ENCODING_* with SawyerEncoding::*

+++ src/openrct2/rct12/SawyerChunkWriter.cpp
@@ -31,7 +31,7 @@ namespace OpenRCT2
     void SawyerChunkWriter::WriteChunk(const void* src, size_t length, SawyerEncoding encoding)
     {
         SawyerCoding::ChunkHeader header;
-        header.encoding = static_cast<uint8_t>(encoding);
+        header.encoding = encoding;
         header.length = static_cast<uint32_t>(length);
 
         auto data = std::make_unique<uint8_t[]>(MAX_COMPRESSED_CHUNK_SIZE);
